Rename project to `dbt-airflow-factory`
