+++ .github/workflows/prepare-release.yml
@@ -15,7 +15,7 @@ jobs:
       matrix:
         python-version: [3.9]
     env:
-      PYTHON_PACKAGE: dbt_airflow_manifest_parser
+      PYTHON_PACKAGE: dbt_airflow_factory
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
