+++ .isort.cfg
@@ -1,6 +1,6 @@
 [settings]
 profile = black
-known_first_party = dbt_airflow_manifest_parser
+known_first_party = dbt_airflow_factory
 default_section = THIRDPARTY
 
 [settings]
