+++ setup.cfg
@@ -3,7 +3,7 @@ current_version = 0.17.0
 
 [bumpversion:file:setup.py]
 
-[bumpversion:file:dbt_airflow_manifest_parser/__init__.py]
+[bumpversion:file:dbt_airflow_factory/__init__.py]
 
 [flake8]
 exclude = .git,__pycache__,build,dist,docs/source/conf.py
