+++ tox.ini
@@ -5,7 +5,7 @@ envlist = py37,py39
 extras =
     tests
 commands=
-    python -m pytest --cov dbt_airflow_manifest_parser --cov-report xml --cov-report term-missing --ignore=venv
+    python -m pytest --cov dbt_airflow_factory --cov-report xml --cov-report term-missing --ignore=venv
 
 # Lint
 [flake8]
